Split dataset checks and tests
